fix: numerical stability + GPU sync elimination + perf caching
Numerical stability (4 fixes):
- freq_layer.py: scale computation in fp32 (1e-9 was zero in bf16 -> NaN)
- kalman.py: Kalman gain division in fp32 (1e-8 was zero in bf16)
- generate.py: multinomial probs in fp32 (bf16 softmax may not sum to 1)
- losses.py: parity clamp min=0.0 (negative parity was penalty not reward)

GPU sync elimination (3 fixes):
- hagi_v4.py: pass cqi_mean as tensor to TurboLoop (no .item() during
  forward; single .item() inside TurboLoop where Python int needed)
- loop.py: remove empty_cache() after distillation (PyTorch allocator
  handles memory; empty_cache causes CUDA synchronization)
- loop.py: return foxp2_gates from train_step, reuse in train() loop
  (was computing grad_stats + gates twice per step)

Performance (1 fix):
- freq_layer.py: cache complex weights + phase for inference (avoid
  recomputing w_re_a @ w_re_b every forward when not training)

Architecture (2 improvements):
- hagi_v4.py: per-position CQI gate in bottleneck + rate dematching
  (was scalar cqi_mean_t; now [B,T] CQI modulates gate per-position =
  per-subband adaptive coding rate, matching 5G NR CQI)
- loop.py: confidence computation in fp32 (avoids bf16 softmax precision
  loss over 49K vocab)

Dead code:
- spectral_cache.py: remove unused advance() method

+++ src/hagi_v4/inference/generate.py
@@ -137,7 +137,7 @@ def generate(
             lt = torch.where(lt < v[..., -1:], float("-inf"), lt)
 
         if temperature > 0:
-            probs = F.softmax(lt, dim=-1)
+            probs = F.softmax(lt.float(), dim=-1)
             return torch.multinomial(probs.reshape(-1, V), 1).reshape(B, n_block)
         else:
             return lt.argmax(dim=-1)
